+++ esp32_arm/src/MyApps/others/acupuncture_2023/touchpad_node.h
@@ -13,7 +13,7 @@ class TouchPad_Node{
 
         String GetMqttPayloadString();
         String GetChannelsPayloadString();
-        int GetSensorValue(int channel_id){ return __all_channels[channel_id].GetSensorValue();};
+        uint8_t GetSensorValue(int channel_id){ return __all_channels[channel_id].GetSensorValue();};
 
         bool __has_changed_channel;  //??
